requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.10.5
requests==2.32.3
python-dotenv==1.0.0
